requirements.txt
pyusb==1.0.2
requests==2.22.0
pycrypto==2.6.1